Update regular expression for new version
Overnight our build of the GPGSuite failed saying:

       Processor: URLTextSearcher: Error: No match found on URL

This is because there has been a new release of the suite (GPG Suite
2016.07 July 4th, 2016 according to the release notes which has version number:

       2016.07_v2

so the regular expression:

    href=\"https://releases.gpgtools.org/GPG_Suite-([0-9a-z\.\-]+)\.dmg\"

was not matching (as it didn't include the underscore character).

This patch just adds the underscore character to the regular expression,
meaning the download now succeeds.